Updating starter code and assignment
