--- symbol_table/src/symbol_table_statement.rs
@@ -1,16 +1,0 @@
-use common::{
-    ast::{
-        ast_struct::ASTNode, data_type::DataType, syntax_element::SyntaxElement
-    }, 
-    error::ErrorType
-};
-
-use crate::symbol_table_struct::{SymbolInfo, SymbolTable, SymbolTableStack, SymbolValue};
-
-impl SymbolTableStack {
-    /// Adds a new variable into the current scope
-    pub fn sym_table_init(&mut self, init_node: &ASTNode) -> Result<(), Vec<ErrorType>> {
-
-    }    
-
-}
